It will open the file target file
It will open the file 'EXCEL1.xlsx' located at 'C:\MyFolder' and will copy the data from all its sheets to the first sheet in the active workbook. The target file will be closed after the data is copied and the code will not save any changes made to it. If you want to keep the changes you should use the "save" method after closing it.